Move some lib tests from test_misc to test_storage.
